重新架构 实现切歌 添加歌曲到列表功能

+++ app/src/main/java/com/example/huangzhangmusic/domain/SongUrlData.kt
@@ -22,7 +22,7 @@ data class Data(
     val level: String,
     val md5: String,
     val payed: Int,
-    val peak: Int,
+    val peak: Float,
     val podcastCtrp: Any,
     val rightSource: Int,
     val size: Int,
